requirements.txt
fastapi
uvicorn[standard]
selenium
undetected-chromedriver
webdriver-manager
python-multipart
pydantic
gspread
google-auth
python-dotenv
